q3/requirements.txt

# Manufacturing Optimization Project - Essential Requirements
# Genetic Algorithm & MIP Solver for Resource Allocation

# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
seaborn
scipy

# Data Visualization
matplotlib>=3.4.0

# HTTP Requests (for OR-Library data download)
requests>=2.25.0

# Mathematical Optimization (MIP Solver)
pulp>=2.6.0

pyyaml>=6.0.0